Added feet and inches as an input option in the edit profile view. This is converted into CM to be stored in the DB. When getting the height from the DB, the conversion is made back into feet/inches for the profile screen

+++ Models/UserProfile.cs
@@ -22,7 +22,7 @@ namespace BodyCompositionCalculator.Models
         [Required]
         public DateTime DateOfBirth { get; set; }
         [Required]
-        public int HeightInCm { get; set; }
+        public double HeightInCm { get; set; }
 
         //Create a foreign key to link back to the user
         //This foreign key will be populated with the ApplicationUser Id Pk column value when a new user is submitted by getting the active user id
